refactor: put the app comonent in its own folder as to give better documentation
