+++ src/app/App.js
@@ -1,0 +1,15 @@
+import React from 'react';
+
+import './App.css';
+
+function App() {
+  return (
+    <div className="App">
+      <header className="App-header">
+        In app container
+      </header>
+    </div>
+  );
+}
+
+export default App;
